requirements.txt
tree-sitter-python==0.23.6
tree-sitter==0.24.0
neo4j==5.28.1
langchain-experimental==0.3.4
langchain-google-genai==2.1.6
py2neo==2021.2.4 
pyvis==0.3.2
streamlit==1.46.1
networkx==3.5